perf: don't set filetype on otter buffers
instead, lsp clients that are configured are attached manually
and the otter nvim-cmp sourced is activated on attach.
This prevents other plugins, vim syntax highlighting and treesitter
from unnecessarily attaching to the otter buffer and significantly
speeds up starting.

TODO: Howver, this now leads to multiple servers for those that attach
automatically on ft. Need to start the server **not** in single file
mode